\name{MBESS}
\alias{MBESS}
\alias{mbess}
\alias{MBES}
\alias{mbes}
\docType{package}
\title{MBESS}
\description{
\code{MBESS} Implements methods that useful in designing research and analyzing data, with 
	particular emphasis on methods that are developed for or used within the behavioral, 
	educational, and social sciences (broadly defined). That being said, many of the methods 
	implemented within MBESS are applicable to a wide variety of disciplines. MBESS has a 
	suite of functions for a variety of related topics, such as effect sizes, confidence intervals 
	for effect sizes (including standardized effect sizes and noncentral effect sizes), sample size
	planning (from the accuracy in parameter estimation [AIPE], power analytic, equivalence, and 
	minimum-risk point estimation perspectives), mediation analysis, various properties of 
	distributions, and a variety of utility functions. MBESS (pronounced `em-bes') was originally 
	an acronym for "Methods for the Behavioral, Educational, and Social Sciences," but at this 
	point MBESS contains methods applicable and used in a wide variety of fields and is an 
	orphan acronym, in the sense that what was an acronym is now literally its name. MBESS has 
	greatly benefited from others, see <http://nd.edu/~kkelley/site/MBESS.html> for a detailed 
	list of those that have contributed and other details.
}
\details{
\tabular{ll}{
Package: \tab MBESS\cr
Type: \tab Package\cr
Version: \tab 4.8.0 \cr
Date: \tab 2020-08-04\cr
License: \tab GPL(>=2)\cr
}
Please read the manual and visit the corresponding web site 
\url{http://nd.edu/~kkelley/site/MBESS.html} 
for information on the capabilities of the \code{MBESS} package. Feel free 
to contact me if there is a feature you would like to see added if it would 
complement the goals of the MBESS package. Beginning with version 
4.8.0, the package also has a home on GitHub \url{https://github.com/yelleKneK/MBESS}. 
Over the years, multiple people have contributed functions to the package. See individual 
functions for details. 
}
\author{
Ken Kelley <\email{KKelley@ND.Edu}; http://www.nd.edu/~kkelley> 

Maintainer: Ken Kelley <\email{KKelley@ND.Edu}; http://www.nd.edu/~kkelley>
}

\keyword{package}
