\name{prof.salary}
\alias{prof.salary}
\docType{data}
\title{Cohen et. al. (2003)'s professor salary data set}
\description{
The data set of the salaries and other information of 62 some professors in Cohen et. al. (2003, pp. 81-82).
}
\usage{data(prof.salary)}
\format{
  A data frame with 62 observations on the following 6 variables.
  \describe{
    \item{\code{id}}{the identification number}
    \item{\code{time}}{the time since getting the Ph.D. degree}
    \item{\code{pub}}{the number of publications}
    \item{\code{sex}}{the gender, 1 for female and 0 for male}
    \item{\code{citation}}{the citation count}
    \item{\code{salary}}{the professor's current salary}
  }
}
%\details{}

\source{
Cohen, J., Cohen, P., West, S. G., & Aiken, L. S. (2003). \emph{Applied multiple regression/correlation analysis for the behavioral sciences} (3rd ed.). Mahwah, NJ: Erlbaum.
}
\references{
Cohen, J., Cohen, P., West, S. G., & Aiken, L. S. (2003). \emph{Applied multiple regression/correlation analysis for the behavioral sciences} (3rd ed.). Mahwah, NJ: Erlbaum.
}
\examples{
data(prof.salary)

}
\keyword{datasets}
